updated hero section and splitted sections
